refactor: remove `ContentLoader` and `unixfs` from `iroh-resolver` package (#566)
- remove `iroh-unixfs` from `iroh-resolver`
- pulls out as much shared structs/enums/traits from `iroh-resolver` as possible and into `iroh-unixfs::types`
- move `content-loader` into `iroh-unixfs`
- adjust indexer test to use more available cid
