+++ Cargo.toml
@@ -14,6 +14,7 @@ members = [
   "iroh-rpc-client",
   "iroh-rpc-types",
   "iroh-share",
+  "iroh-unixfs",
   "iroh-store",
   "iroh-util",
   "stores/*",
