+++ examples/Cargo.toml
@@ -17,7 +17,7 @@ clap = { version = "4.0.9", features = ["derive"] }
 futures = "0.3.21"
 indicatif = "0.17.0"
 iroh-car = { version = "0.1.3", path = "../iroh-car" }
-iroh-resolver = { version = "0.1.3", path = "../iroh-resolver" }
+iroh-unixfs = { version = "0.1.3", path = "../iroh-unixfs" }
 iroh-rpc-client = { version = "0.1.3", path = "../iroh-rpc-client" }
 iroh-util = { version = "0.1.3", path = "../iroh-util" }
 par-stream = { version = "0.10.2", default-features = false, features = ["runtime-tokio"] }
