+++ iroh-api/Cargo.toml
@@ -22,6 +22,7 @@ iroh-metrics = { version = "0.1.3", path = "../iroh-metrics", default-features =
 iroh-resolver = { version = "0.1.3", path = "../iroh-resolver" }
 iroh-rpc-client = { version = "0.1.3", path = "../iroh-rpc-client", default-features = false }
 iroh-rpc-types = { version = "0.1.3", path = "../iroh-rpc-types", default-features = false }
+iroh-unixfs = { version = "0.1.3", path = "../iroh-unixfs" }
 iroh-util = { version = "0.1.3", path = "../iroh-util" }
 libp2p = "0.50"
 mockall = { version = "0.11.2", optional = true }
