+++ iroh-api/src/lib.rs
@@ -18,9 +18,9 @@ pub use crate::p2p::P2p as P2pApi;
 pub use crate::p2p::PeerIdOrAddr;
 pub use bytes::Bytes;
 pub use cid::Cid;
-pub use iroh_resolver::chunker::DEFAULT_CHUNKS_SIZE;
 pub use iroh_resolver::resolver::Path as IpfsPath;
-pub use iroh_resolver::unixfs_builder::{AddEvent, ChunkerConfig, Config as UnixfsConfig};
 pub use iroh_rpc_client::{Lookup, ServiceStatus, StatusRow, StatusTable};
+pub use iroh_unixfs::builder::{AddEvent, Config as UnixfsConfig};
+pub use iroh_unixfs::chunker::{ChunkerConfig, DEFAULT_CHUNKS_SIZE};
 pub use libp2p::gossipsub::MessageId;
 pub use libp2p::{Multiaddr, PeerId};
