+++ iroh-one/src/uds.rs
@@ -5,7 +5,7 @@ use axum::{Router, Server};
 use futures::ready;
 use hyper::server::accept::Accept;
 use iroh_gateway::{core::State, handlers::get_app_routes};
-use iroh_resolver::content_loader::ContentLoader;
+use iroh_unixfs::content_loader::ContentLoader;
 use std::path::PathBuf;
 use std::{
     pin::Pin,
