+++ iroh-resolver/src/lib.rs
@@ -1,12 +1,4 @@
-pub mod balanced_tree;
-pub mod chunker;
-pub mod codecs;
-pub mod content_loader;
 pub mod dns_resolver;
-pub mod hamt;
-pub mod indexer;
 pub mod resolver;
-pub mod unixfs;
-pub mod unixfs_builder;
 
-pub use crate::resolver::parse_links;
+pub use resolver::{Path, PathType};
