+++ iroh-unixfs/src/builder.rs
@@ -1,12 +1,11 @@
 use std::{
-    fmt::{Debug, Display},
+    fmt::Debug,
     path::{Path, PathBuf},
     pin::Pin,
-    str::FromStr,
     sync::Arc,
 };
 
-use anyhow::{anyhow, ensure, Context, Result};
+use anyhow::{ensure, Result};
 use async_recursion::async_recursion;
 use async_trait::async_trait;
 use bytes::Bytes;
@@ -19,8 +18,8 @@ use tokio::io::AsyncRead;
 
 use crate::{
     balanced_tree::{TreeBuilder, DEFAULT_DEGREE},
-    chunker::{self, Chunker, DEFAULT_CHUNKS_SIZE, DEFAULT_CHUNK_SIZE_LIMIT},
-    resolver::Block,
+    chunker::{self, Chunker, ChunkerConfig, DEFAULT_CHUNK_SIZE_LIMIT},
+    types::Block,
     unixfs::{dag_pb, unixfs_pb, DataType, Node, UnixfsNode},
 };
 
@@ -592,60 +591,6 @@ pub struct Config {
     pub chunker: ChunkerConfig,
 }
 
-/// Chunker configuration.
-#[derive(Debug, Clone, PartialEq, Eq, Copy)]
-pub enum ChunkerConfig {
-    /// Fixed sized chunker.
-    Fixed(usize),
-    /// Rabin chunker.
-    Rabin,
-}
-
-impl Display for ChunkerConfig {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Fixed(chunk_size) => write!(f, "fixed-{}", chunk_size),
-            Self::Rabin => write!(f, "rabin"),
-        }
-    }
-}
-
-impl FromStr for ChunkerConfig {
-    type Err = anyhow::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        if s == "rabin" {
-            return Ok(ChunkerConfig::Rabin);
-        }
-
-        if let Some(rest) = s.strip_prefix("fixed") {
-            if rest.is_empty() {
-                return Ok(ChunkerConfig::Fixed(DEFAULT_CHUNKS_SIZE));
-            }
-
-            if let Some(rest) = rest.strip_prefix('-') {
-                let chunk_size: usize = rest.parse().context("invalid chunk size")?;
-                if chunk_size > DEFAULT_CHUNK_SIZE_LIMIT {
-                    return Err(anyhow!("chunk size too large"));
-                }
-
-                return Ok(ChunkerConfig::Fixed(chunk_size));
-            }
-        }
-
-        Err(anyhow!("unknown chunker: {}", s))
-    }
-}
-
-impl From<ChunkerConfig> for Chunker {
-    fn from(cfg: ChunkerConfig) -> Self {
-        match cfg {
-            ChunkerConfig::Fixed(chunk_size) => Chunker::Fixed(chunker::Fixed::new(chunk_size)),
-            ChunkerConfig::Rabin => Chunker::Rabin(Box::new(chunker::Rabin::default())),
-        }
-    }
-}
-
 /// Adds a single file.
 /// - storing the content using `rpc.store`
 /// - returns a stream of AddEvent
@@ -830,55 +775,11 @@ async fn make_dir_from_path<P: Into<PathBuf>>(path: P, chunker: Chunker) -> Resu
     dir.build()
 }
 
-/// Read an `AsyncRead` into a `Vec` completely.
-#[doc(hidden)]
-pub async fn read_to_vec<T: AsyncRead + Unpin>(mut reader: T) -> Result<Vec<u8>> {
-    use tokio::io::AsyncReadExt;
-
-    let mut out = Vec::new();
-    reader.read_to_end(&mut out).await?;
-    Ok(out)
-}
-
-/// Read a stream of (cid, block) pairs into an in memory store and return the store and the root cid.
-#[doc(hidden)]
-pub async fn stream_to_resolver(
-    stream: impl Stream<Item = Result<Block>>,
-) -> Result<(
-    Cid,
-    crate::resolver::Resolver<Arc<fnv::FnvHashMap<Cid, Bytes>>>,
-)> {
-    tokio::pin!(stream);
-    let blocks: Vec<_> = stream.try_collect().await?;
-    for block in &blocks {
-        block.validate()?;
-    }
-    let root_block = blocks.last().context("no root")?.clone();
-    let store: fnv::FnvHashMap<Cid, Bytes> = blocks
-        .into_iter()
-        .map(|block| {
-            let (cid, bytes, _) = block.into_parts();
-            (cid, bytes)
-        })
-        .collect();
-    let resolver = crate::resolver::Resolver::new(Arc::new(store));
-    Ok((*root_block.cid(), resolver))
-}
-
 #[cfg(test)]
 mod tests {
-    use crate::chunker::DEFAULT_CHUNKS_SIZE;
-    use crate::content_loader::ContentLoader;
-    use crate::resolver::{Out, OutMetrics, Resolver, ResponseClip};
-
     use super::*;
-    use anyhow::Result;
-    use futures::TryStreamExt;
-    use proptest::prelude::*;
-    use rand::prelude::*;
-    use rand_chacha::ChaCha8Rng;
-    use std::{collections::BTreeMap, io::prelude::*};
-    use tokio::io::AsyncReadExt;
+    use crate::chunker::DEFAULT_CHUNKS_SIZE;
+    use std::io::Write;
 
     #[tokio::test]
     async fn test_builder_basics() -> Result<()> {
@@ -993,161 +894,6 @@ mod tests {
         // TODO: check content
         Ok(())
     }
-
-    #[derive(Debug, Clone, PartialEq, Eq)]
-    enum TestDirEntry {
-        File(Bytes),
-        Directory(TestDir),
-    }
-    type TestDir = BTreeMap<String, TestDirEntry>;
-
-    /// builds an unixfs directory out of a TestDir
-    #[async_recursion(?Send)]
-    async fn build_directory(name: &str, dir: &TestDir) -> Result<Directory> {
-        let mut builder = DirectoryBuilder::new();
-        builder.name(name);
-        for (name, entry) in dir {
-            match entry {
-                TestDirEntry::File(content) => {
-                    let file = FileBuilder::new()
-                        .name(name)
-                        .content_bytes(content.to_vec())
-                        .build()
-                        .await?;
-                    builder.add_file(file);
-                }
-                TestDirEntry::Directory(dir) => {
-                    let dir = build_directory(name, dir).await?;
-                    builder.add_dir(dir)?;
-                }
-            }
-        }
-        builder.build()
-    }
-
-    /// builds a TestDir out of a stream of blocks and a resolver
-    async fn build_testdir(
-        stream: impl Stream<Item = Result<(crate::resolver::Path, Out)>>,
-        resolver: Resolver<impl ContentLoader + Unpin>,
-    ) -> Result<TestDir> {
-        tokio::pin!(stream);
-
-        /// recursively create directories for a path
-        fn mkdir(dir: &mut TestDir, path: &[String]) -> Result<()> {
-            if let Some((first, rest)) = path.split_first() {
-                if let TestDirEntry::Directory(child) = dir
-                    .entry(first.clone())
-                    .or_insert_with(|| TestDirEntry::Directory(Default::default()))
-                {
-                    mkdir(child, rest)?;
-                } else {
-                    anyhow::bail!("not a directory");
-                }
-            }
-            Ok(())
-        }
-
-        /// create a file in a directory hierarchy
-        fn mkfile(dir: &mut TestDir, path: &[String], data: Bytes) -> Result<()> {
-            if let Some((first, rest)) = path.split_first() {
-                if rest.is_empty() {
-                    dir.insert(first.clone(), TestDirEntry::File(data));
-                } else if let TestDirEntry::Directory(child) = dir
-                    .entry(first.clone())
-                    .or_insert_with(|| TestDirEntry::Directory(Default::default()))
-                {
-                    mkfile(child, rest, data)?;
-                } else {
-                    anyhow::bail!("not a directory");
-                }
-            }
-            Ok(())
-        }
-
-        let reference = stream
-            .try_fold(TestDir::default(), move |mut agg, (path, item)| {
-                let resolver = resolver.clone();
-                async move {
-                    if item.is_dir() {
-                        mkdir(&mut agg, path.tail())?;
-                    } else {
-                        let reader = item.pretty(
-                            resolver.clone(),
-                            OutMetrics::default(),
-                            ResponseClip::NoClip,
-                        )?;
-                        let data = read_to_vec(reader).await?;
-                        mkfile(&mut agg, path.tail(), data.into())?;
-                    }
-                    Ok(agg)
-                }
-            })
-            .await?;
-        Ok(reference)
-    }
-
-    /// a roundtrip test that converts a dir to an unixfs DAG and back
-    async fn dir_roundtrip_test(dir: TestDir) -> Result<bool> {
-        let directory = build_directory("", &dir).await?;
-        let stream = directory.encode();
-        let (root, resolver) = stream_to_resolver(stream).await?;
-        let stream = resolver.resolve_recursive_with_paths(crate::resolver::Path::from_cid(root));
-        let reference = build_testdir(stream, resolver).await?;
-        Ok(dir == reference)
-    }
-
-    /// sync version of dir_roundtrip_test for use in proptest
-    fn dir_roundtrip_test_sync(dir: TestDir) -> bool {
-        tokio::runtime::Builder::new_current_thread()
-            .build()
-            .unwrap()
-            .block_on(dir_roundtrip_test(dir))
-            .unwrap()
-    }
-
-    /// a roundtrip test that converts a file to an unixfs DAG and back
-    async fn file_roundtrip_test(data: Bytes, chunk_size: usize, degree: usize) -> Result<bool> {
-        let file = FileBuilder::new()
-            .name("file.bin")
-            .fixed_chunker(chunk_size)
-            .degree(degree)
-            .content_bytes(data.clone())
-            .build()
-            .await?;
-        let stream = file.encode().await?;
-        let (root, resolver) = stream_to_resolver(stream).await?;
-        let out = resolver
-            .resolve(crate::resolver::Path::from_cid(root))
-            .await?;
-        let t =
-            read_to_vec(out.pretty(resolver, OutMetrics::default(), ResponseClip::NoClip)?).await?;
-        println!("{}", data.len());
-        Ok(t == data)
-    }
-
-    /// a roundtrip test that converts a symlink to a unixfs DAG and back
-    #[tokio::test]
-    async fn symlink_roundtrip_test() -> Result<()> {
-        let mut builder = SymlinkBuilder::new("foo");
-        let target = "../../bar.txt";
-        builder.target(target);
-        let sym = builder.build().await?;
-        let block = sym.encode()?;
-        let stream = async_stream::try_stream! {
-            yield block;
-        };
-        let (root, resolver) = stream_to_resolver(stream).await?;
-        let out = resolver
-            .resolve(crate::resolver::Path::from_cid(root))
-            .await?;
-        let mut reader = out.pretty(resolver, OutMetrics::default(), ResponseClip::NoClip)?;
-        let mut t = String::new();
-        reader.read_to_string(&mut t).await?;
-        println!("{}", t);
-        assert_eq!(target, t);
-        Ok(())
-    }
-
     #[cfg(not(windows))]
     #[tokio::test]
     async fn symlink_from_disk_test() -> Result<()> {
@@ -1163,70 +909,6 @@ mod tests {
         assert_eq!(expect_target, got_symlink.target);
         Ok(())
     }
-
-    /// sync version of file_roundtrip_test for use in proptest
-    fn file_roundtrip_test_sync(data: Bytes, chunk_size: usize, degree: usize) -> bool {
-        let f = file_roundtrip_test(data, chunk_size, degree);
-        tokio::runtime::Builder::new_current_thread()
-            .build()
-            .unwrap()
-            .block_on(f)
-            .unwrap()
-    }
-
-    fn arb_test_dir() -> impl Strategy<Value = TestDir> {
-        // create an arbitrary nested directory structure
-        fn arb_dir_entry() -> impl Strategy<Value = TestDirEntry> {
-            let leaf = any::<Vec<u8>>().prop_map(|x| TestDirEntry::File(Bytes::from(x)));
-            leaf.prop_recursive(3, 64, 10, |inner| {
-                prop::collection::btree_map(".*", inner, 0..10).prop_map(TestDirEntry::Directory)
-            })
-        }
-        prop::collection::btree_map(".*", arb_dir_entry(), 0..10)
-    }
-
-    fn arb_degree() -> impl Strategy<Value = usize> {
-        // use either the smallest possible degree for complex tree structures, or the default value for realism
-        prop_oneof![Just(2), Just(DEFAULT_DEGREE)]
-    }
-
-    fn arb_chunk_size() -> impl Strategy<Value = usize> {
-        // use either the smallest possible chunk size for complex tree structures, or the default value for realism
-        prop_oneof![Just(1), Just(DEFAULT_CHUNKS_SIZE)]
-    }
-
-    proptest! {
-        #[test]
-        fn test_file_roundtrip(data in proptest::collection::vec(any::<u8>(), 0usize..1024), chunk_size in arb_chunk_size(), degree in arb_degree()) {
-            assert!(file_roundtrip_test_sync(data.into(), chunk_size, degree));
-        }
-
-        #[test]
-        fn test_dir_roundtrip(data in arb_test_dir()) {
-            assert!(dir_roundtrip_test_sync(data));
-        }
-    }
-
-    #[tokio::test]
-    async fn test_builder_roundtrip_complex_tree_1() -> Result<()> {
-        // fill with random data so we get distinct cids for all blocks
-        let mut rng = ChaCha8Rng::from_seed([0; 32]);
-        let mut data = vec![0u8; 1024 * 128];
-        rng.fill(data.as_mut_slice());
-        assert!(file_roundtrip_test(data.into(), 1024, 4).await?);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_builder_roundtrip_128m() -> Result<()> {
-        // fill with random data so we get distinct cids for all blocks
-        let mut rng = ChaCha8Rng::from_seed([0; 32]);
-        let mut data = vec![0u8; 128 * 1024 * 1024];
-        rng.fill(data.as_mut_slice());
-        assert!(file_roundtrip_test(data.into(), DEFAULT_CHUNKS_SIZE, DEFAULT_DEGREE).await?);
-        Ok(())
-    }
-
     #[tokio::test]
     async fn test_builder_stream_large() -> Result<()> {
         // Create a directory
