+++ iroh-unixfs/src/unixfs.rs
@@ -1,7 +1,6 @@
 use std::{
     collections::VecDeque,
-    fmt::{self, Debug},
-    io::Cursor,
+    fmt::Debug,
     pin::Pin,
     task::{Context, Poll},
 };
@@ -10,15 +9,16 @@ use anyhow::{anyhow, bail, ensure, Result};
 use bytes::{Buf, Bytes};
 use cid::{multihash::MultihashDigest, Cid};
 use futures::{future::BoxFuture, stream::BoxStream, FutureExt, Stream, StreamExt};
+use iroh_metrics::resolver::OutMetrics;
 use prost::Message;
 use tokio::io::{AsyncRead, AsyncSeek};
 
 use crate::{
     chunker::DEFAULT_CHUNK_SIZE_LIMIT,
     codecs::Codec,
-    content_loader::ContentLoader,
+    content_loader::{ContentLoader, LoaderContext},
     hamt::Hamt,
-    resolver::{Block, LoaderContext, OutMetrics, Resolver, ResponseClip},
+    types::{Block, Link, LinkRef, Links, PbLinks, ResponseClip},
 };
 
 pub(crate) mod unixfs_pb {
@@ -62,40 +62,6 @@ impl Unixfs {
 
     pub fn data(&self) -> Option<&Bytes> {
         self.inner.data.as_ref()
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Link {
-    pub cid: Cid,
-    pub name: Option<String>,
-    pub tsize: Option<u64>,
-}
-
-impl Link {
-    pub fn as_ref(&self) -> LinkRef<'_> {
-        LinkRef {
-            cid: self.cid,
-            name: self.name.as_deref(),
-            tsize: self.tsize,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct LinkRef<'a> {
-    pub cid: Cid,
-    pub name: Option<&'a str>,
-    pub tsize: Option<u64>,
-}
-
-impl LinkRef<'_> {
-    pub fn to_owned(&self) -> Link {
-        Link {
-            cid: self.cid,
-            name: self.name.map(|t| t.to_string()),
-            tsize: self.tsize,
-        }
     }
 }
 
@@ -340,10 +306,10 @@ impl UnixfsNode {
     }
 
     /// If this is a directory or hamt shard, returns a stream that yields all children of it.
-    pub fn as_child_reader<'a, 'b: 'a, T: ContentLoader>(
+    pub fn as_child_reader<'a, 'b: 'a, C: ContentLoader>(
         &'a self,
         ctx: LoaderContext,
-        loader: &'b Resolver<T>,
+        loader: C,
         om: OutMetrics,
     ) -> Result<Option<UnixfsChildStream<'a>>> {
         match self {
@@ -368,13 +334,13 @@ impl UnixfsNode {
         }
     }
 
-    pub fn into_content_reader<T: ContentLoader>(
+    pub fn into_content_reader<C: ContentLoader>(
         self,
         ctx: LoaderContext,
-        loader: Resolver<T>,
+        loader: C,
         om: OutMetrics,
         pos_max: ResponseClip,
-    ) -> Result<Option<UnixfsContentReader<T>>> {
+    ) -> Result<Option<UnixfsContentReader<C>>> {
         match self {
             UnixfsNode::Raw(_)
             | UnixfsNode::RawNode(_)
@@ -411,33 +377,21 @@ pub enum UnixfsChildStream<'a> {
     },
 }
 
-impl<'a> fmt::Debug for UnixfsChildStream<'a> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+impl<'a> Debug for UnixfsChildStream<'a> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
-            Self::Hamt {
-                stream: _,
-                pos,
-                out_metrics,
-            } => f
-                .debug_struct("Hamt")
-                .field("stream", &"impl Stream<Item=Result<Link>>")
-                .field("pos", pos)
-                .field("out_metrics", out_metrics)
-                .finish(),
-            Self::Directory {
-                stream: _,
-                out_metrics,
-            } => f
-                .debug_struct("Directory")
-                .field("stream", &"impl Stream<Item=Result<Link>")
-                .field("out_metrics", out_metrics)
-                .finish(),
+            UnixfsChildStream::Hamt {
+                pos, out_metrics, ..
+            } =>
+                write!(f, "UnixfsChildStream::Hamt {{ stream: BoxStream<Result<Link>>, pos: {}, out_metrics {:?} }}", pos, out_metrics),
+            UnixfsChildStream::Directory { out_metrics, .. } =>
+                write!(f, "UnixfsChildStream::Directory {{ stream: BoxStream<Result<Link>>, out_metrics {:?} }}", out_metrics),
         }
     }
 }
 
 #[derive(Debug)]
-pub enum UnixfsContentReader<T: ContentLoader> {
+pub enum UnixfsContentReader<C: ContentLoader> {
     File {
         root_node: UnixfsNode,
         /// Absolute position in bytes
@@ -450,13 +404,13 @@ pub enum UnixfsContentReader<T: ContentLoader> {
         current_node: CurrentNodeState,
         /// Stack of links left to traverse.
         current_links: Vec<VecDeque<Link>>,
-        loader: Resolver<T>,
+        loader: C,
         out_metrics: OutMetrics,
         ctx: std::sync::Arc<tokio::sync::Mutex<LoaderContext>>,
     },
 }
 
-impl<T: ContentLoader> UnixfsContentReader<T> {
+impl<C: ContentLoader> UnixfsContentReader<C> {
     /// Returns the size in bytes, if known in advance.
     pub fn size(&self) -> Option<u64> {
         match self {
@@ -486,7 +440,7 @@ impl Stream for UnixfsChildStream<'_> {
     }
 }
 
-impl<T: ContentLoader + Unpin + 'static> AsyncRead for UnixfsContentReader<T> {
+impl<C: ContentLoader + Unpin + 'static> AsyncRead for UnixfsContentReader<C> {
     fn poll_read(
         mut self: Pin<&mut Self>,
         cx: &mut Context<'_>,
@@ -541,7 +495,7 @@ impl<T: ContentLoader + Unpin + 'static> AsyncRead for UnixfsContentReader<T> {
     }
 }
 
-impl<T: ContentLoader + Unpin + 'static> AsyncSeek for UnixfsContentReader<T> {
+impl<C: ContentLoader + Unpin + 'static> AsyncSeek for UnixfsContentReader<C> {
     fn start_seek(mut self: Pin<&mut Self>, position: std::io::SeekFrom) -> std::io::Result<()> {
         match &mut *self {
             UnixfsContentReader::File {
@@ -671,10 +625,10 @@ impl Debug for CurrentNodeState {
     }
 }
 
-fn load_next_node<T: ContentLoader + 'static>(
+fn load_next_node<C: ContentLoader + 'static>(
     current_node: &mut CurrentNodeState,
     current_links: &mut Vec<VecDeque<Link>>,
-    loader: Resolver<T>,
+    loader: C,
     ctx: std::sync::Arc<tokio::sync::Mutex<LoaderContext>>,
 ) -> bool {
     // Load next node
@@ -699,7 +653,7 @@ fn load_next_node<T: ContentLoader + 'static>(
 
     let fut = async move {
         let ctx = ctx.lock().await;
-        let loaded_cid = loader.loader().load_cid(&link.cid, &ctx).await?;
+        let loaded_cid = loader.load_cid(&link.cid, &ctx).await?;
         let node = UnixfsNode::decode(&link.cid, loaded_cid.data)?;
 
         Ok(node)
@@ -710,10 +664,10 @@ fn load_next_node<T: ContentLoader + 'static>(
 }
 
 #[allow(clippy::too_many_arguments)]
-fn poll_read_file_at<T: ContentLoader + 'static>(
+fn poll_read_file_at<C: ContentLoader + 'static>(
     cx: &mut Context<'_>,
     root_node: &Node,
-    loader: Resolver<T>,
+    loader: C,
     pos: &mut usize,
     skip_pos: &mut usize,
     pos_max: ResponseClip,
@@ -873,84 +827,5 @@ fn poll_read_file_at<T: ContentLoader + 'static>(
                 }
             }
         }
-    }
-}
-
-#[derive(Debug)]
-pub enum Links<'a> {
-    Raw,
-    RawNode(PbLinks<'a>),
-    Directory(PbLinks<'a>),
-    File(PbLinks<'a>),
-    Symlink(PbLinks<'a>),
-    HamtShard(PbLinks<'a>),
-}
-
-#[derive(Debug)]
-pub struct PbLinks<'a> {
-    i: usize,
-    outer: &'a dag_pb::PbNode,
-}
-
-impl<'a> PbLinks<'a> {
-    pub fn new(outer: &'a dag_pb::PbNode) -> Self {
-        PbLinks { i: 0, outer }
-    }
-}
-
-impl<'a> Iterator for Links<'a> {
-    type Item = Result<LinkRef<'a>>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        match self {
-            Links::Raw => None,
-            Links::Directory(links)
-            | Links::RawNode(links)
-            | Links::File(links)
-            | Links::Symlink(links)
-            | Links::HamtShard(links) => links.next(),
-        }
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        match self {
-            Links::Raw => (0, Some(0)),
-            Links::Directory(links)
-            | Links::RawNode(links)
-            | Links::File(links)
-            | Links::Symlink(links)
-            | Links::HamtShard(links) => links.size_hint(),
-        }
-    }
-}
-
-impl<'a> Iterator for PbLinks<'a> {
-    type Item = Result<LinkRef<'a>>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.i == self.outer.links.len() {
-            return None;
-        }
-
-        let l = &self.outer.links[self.i];
-        self.i += 1;
-
-        let res = l
-            .hash
-            .as_ref()
-            .ok_or_else(|| anyhow!("missing link"))
-            .and_then(|c| {
-                Ok(LinkRef {
-                    cid: Cid::read_bytes(Cursor::new(c))?,
-                    name: l.name.as_deref(),
-                    tsize: l.tsize,
-                })
-            });
-
-        Some(res)
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        (self.outer.links.len(), Some(self.outer.links.len()))
     }
 }
